Changes to accommodate large numeric values in the input string
Changed data type from int to long to accommodate larger numeric value in the input string

+++ Exercise/src/StringExercise/Main.java
@@ -4,9 +4,9 @@ import java.util.Scanner;
 
 public class Main {
 
-	static int maxNumber(String S) {
-		int maxval = 0;
-		int currentval = 0;
+	static long maxNumber(String S) {
+		long maxval = 0;
+		long currentval = 0;
 
 		if (S == null)
 			return 0;
@@ -19,7 +19,7 @@ public class Main {
 			if (Character.isDigit(S.charAt(i))) {
 				currentval = (currentval * 10) + Character.getNumericValue(S.charAt(i));
 
-				// update the maximum value if end of string is reached 
+				// update the maximum value if end of string is reached
 				// with last character being digit
 				if (i == S.length() - 1) {
 					maxval = Math.max(maxval, currentval);
